Move test setup to mock file
